backend/requirements.txt
fastapi
uvicorn
python-multipart
boto3
openai
pandas
numpy
python-dotenv
torch
transformers
huggingface-hub
malaya
holidays
PySastrawi
websockets
fastapi-cache2